completed task 5 and 6

+++ models/articles.model.js
@@ -12,7 +12,8 @@ function fetchArticleId(article_id) {
 }
 
 function fetchAllArticles() {
-    return db.query(
+    return db
+      .query(
         `SELECT 
     articles.article_id, 
     articles.title, 
@@ -21,7 +22,7 @@ function fetchAllArticles() {
     articles.created_at,
     articles.votes,
     articles.article_img_url,
-    COUNT(comments.comment_id) AS comment_count 
+    COUNT(comments.comment_id)::INTEGER AS comment_count
     FROM articles
     LEFT JOIN comments ON articles.article_id = comments.article_id
     GROUP BY articles.article_id
